Create Fast Sorting Algorithm
Implement a fast sorting algorithm as described in class. Here
is some example code (Links to an external site.) to use as a guideline.
You need only implement the sorting algorithm; the main
method will be provided for you.

You must hand in the Sorts class with the
heapSort sorting algorithm implemented.